corregi el link json js

+++ js/json.js
@@ -14,28 +14,28 @@ const guardoDatosJSON = () => {
 
 const reset = document.getElementById("reset");
 reset.addEventListener("click", () => {
-    //agregue la libreria sweet alert para la confirmacion antes de borrar los datos
-    const resp = Swal.fire({
+//agregue la libreria sweet alert para la confirmacion antes de borrar los datos
+    const resp =Swal.fire({
         title: '¿Realmente desea eliminar los datos almacenados?',
-
+        
         icon: 'warning',
         showCancelButton: true,
         confirmButtonColor: 'grey',
         cancelButtonColor: '#ad2424',
         confirmButtonText: 'Si, borrar todo!'
-    }).then((result) => {
+      }).then((result) => {
         if (result.isConfirmed) {
             localStorage.clear();
             formulario.reset();
-            Swal.fire(
-                'Borrado!',
-                'Tus datos han sido borrado.',
-                'success'
-
-
-            )
+          Swal.fire(
+            'Borrado!',
+            'Tus datos han sido borrado.',
+            'success'
+            
+            
+          )
         }
-    })
+      })
 
 })
